fix(consent): split shared constants from server-only cookie helper
Build was failing because client-cookie-consent.ts imported the constants
and parser from cookie-consent.ts, and that file imports next/headers.
Even though the client file only used the pure constants, bundling pulled
in the server-only module and broke the build:

  ./src/lib/cookie-consent.ts (1:1)
  You're importing a module that depends on "next/headers".

Move the constants, CookieConsentValue type, and parseCookieConsent into
cookie-consent.shared.ts (no runtime deps). cookie-consent.ts now
re-exports them so existing server callers don't break, and adds
'server-only' for explicit boundary safety. client-cookie-consent.ts
imports from the shared module directly.
